+++ scripts/rollback/rollback-3.5.2.sh
@@ -1,0 +1,76 @@
+#!/bin/bash
+# timestamp: 1662654886011
+
+npm dist-tag add @midwayjs/fc-starter@3.5.1 latest
+npm dist-tag add @midwayjs/fc-starter@3.5.2 beta
+npm dist-tag add @midwayjs/serverless-http-parser@3.5.1 latest
+npm dist-tag add @midwayjs/serverless-http-parser@3.5.2 beta
+npm dist-tag add @midwayjs/code-dye@3.5.1 latest
+npm dist-tag add @midwayjs/code-dye@3.5.2 beta
+npm dist-tag add @midwayjs/cross-domain@3.5.1 latest
+npm dist-tag add @midwayjs/cross-domain@3.5.2 beta
+npm dist-tag add @midwayjs/faas@3.5.1 latest
+npm dist-tag add @midwayjs/faas@3.5.2 beta
+npm dist-tag add @midwayjs/http-proxy@3.5.1 latest
+npm dist-tag add @midwayjs/http-proxy@3.5.2 beta
+npm dist-tag add @midwayjs/i18n@3.5.1 latest
+npm dist-tag add @midwayjs/i18n@3.5.2 beta
+npm dist-tag add @midwayjs/info@3.5.1 latest
+npm dist-tag add @midwayjs/info@3.5.2 beta
+npm dist-tag add @midwayjs/security@3.5.1 latest
+npm dist-tag add @midwayjs/security@3.5.2 beta
+npm dist-tag add @midwayjs/static-file@3.5.1 latest
+npm dist-tag add @midwayjs/static-file@3.5.2 beta
+npm dist-tag add @midwayjs/swagger@3.5.1 latest
+npm dist-tag add @midwayjs/swagger@3.5.2 beta
+npm dist-tag add @midwayjs/upload@3.5.1 latest
+npm dist-tag add @midwayjs/upload@3.5.2 beta
+npm dist-tag add @midwayjs/validate@3.5.1 latest
+npm dist-tag add @midwayjs/validate@3.5.2 beta
+npm dist-tag add @midwayjs/version@3.5.1 latest
+npm dist-tag add @midwayjs/version@3.5.2 beta
+tnpm dist-tag add @midwayjs/fc-starter@3.5.1 latest
+tnpm dist-tag add @midwayjs/fc-starter@3.5.2 beta
+tnpm dist-tag add @midwayjs/serverless-http-parser@3.5.1 latest
+tnpm dist-tag add @midwayjs/serverless-http-parser@3.5.2 beta
+tnpm dist-tag add @midwayjs/code-dye@3.5.1 latest
+tnpm dist-tag add @midwayjs/code-dye@3.5.2 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.5.1 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.5.2 beta
+tnpm dist-tag add @midwayjs/faas@3.5.1 latest
+tnpm dist-tag add @midwayjs/faas@3.5.2 beta
+tnpm dist-tag add @midwayjs/http-proxy@3.5.1 latest
+tnpm dist-tag add @midwayjs/http-proxy@3.5.2 beta
+tnpm dist-tag add @midwayjs/i18n@3.5.1 latest
+tnpm dist-tag add @midwayjs/i18n@3.5.2 beta
+tnpm dist-tag add @midwayjs/info@3.5.1 latest
+tnpm dist-tag add @midwayjs/info@3.5.2 beta
+tnpm dist-tag add @midwayjs/security@3.5.1 latest
+tnpm dist-tag add @midwayjs/security@3.5.2 beta
+tnpm dist-tag add @midwayjs/static-file@3.5.1 latest
+tnpm dist-tag add @midwayjs/static-file@3.5.2 beta
+tnpm dist-tag add @midwayjs/swagger@3.5.1 latest
+tnpm dist-tag add @midwayjs/swagger@3.5.2 beta
+tnpm dist-tag add @midwayjs/upload@3.5.1 latest
+tnpm dist-tag add @midwayjs/upload@3.5.2 beta
+tnpm dist-tag add @midwayjs/validate@3.5.1 latest
+tnpm dist-tag add @midwayjs/validate@3.5.2 beta
+tnpm dist-tag add @midwayjs/version@3.5.1 latest
+tnpm dist-tag add @midwayjs/version@3.5.2 beta
+
+# Changes:
+
+#  - @midwayjs/fc-starter: 3.5.1 => 3.5.2
+#  - @midwayjs/serverless-http-parser: 3.5.1 => 3.5.2
+#  - @midwayjs/code-dye: 3.5.1 => 3.5.2
+#  - @midwayjs/cross-domain: 3.5.1 => 3.5.2
+#  - @midwayjs/faas: 3.5.1 => 3.5.2
+#  - @midwayjs/http-proxy: 3.5.1 => 3.5.2
+#  - @midwayjs/i18n: 3.5.1 => 3.5.2
+#  - @midwayjs/info: 3.5.1 => 3.5.2
+#  - @midwayjs/security: 3.5.1 => 3.5.2
+#  - @midwayjs/static-file: 3.5.1 => 3.5.2
+#  - @midwayjs/swagger: 3.5.1 => 3.5.2
+#  - @midwayjs/upload: 3.5.1 => 3.5.2
+#  - @midwayjs/validate: 3.5.1 => 3.5.2
+#  - @midwayjs/version: 3.5.1 => 3.5.2